Refactoring the project structure to remove the "web" folder
The original intend was store all parts of the project in a single GitHub repo but I think it makes more sense to decompose this project into a few repos.  This repo will be renamed to reflect the nature that it will only handle the website itself and nothing related to the NodeMCU work or Alexa integration.  This also helps to simplify some of the hierarchy in this repo too.  Having everything in this project seemed to cause confusion from the outside as to what this repo's goal was--hopefully this will resolve this.  Lastly, it'll have advantages on increasing the granularity of Slack channels for this project as a whole.