Changing the casing of the folder names
This applies to folders inside of the F3 application.  Because of the autoloader for F3, if I want to use class names with capital letters to start, the folder names also need capital letters or else it won't be able to find the classes.  I had changed the naming locally, but Git doesn't pick up on changes in case unless you explicitly use `git mv` to move the folders to a name with a capital letter to start.  This would only be a problem for Mac and Linux operating systems.