Introduced middleware to add "Content-Length" header to a response
Using chunked transfer encoding as the HTTP response was causing issues for the ESP8266 when hitting the `info` endpoint.  The ESP8266 was trying to read the chunk size as part of the response.  I don't need chunked transfer encoding for this, so reverting to "Content-Length" will work and is probably better for limited capability hardware like the ESP8266.