Added default implementation for IDeviceInformation
The `ErrantDeviceInformation` class will always return a 400 error.  This class is used as a default implementation of `IDeviceInformation` to be passed into the API's `DevicesController`.  The reason this is necessary is before this change, calling an endpoint (besides `info(Request $request)`) in the API's `DevicesController` would cause a 500 error because there wasn't a value for `IDeviceInformation` it was left unregistered.  This will now more gracefully give the user a 400 error.  Another option to solve this that I approached was to remove `IDeviceInformation` from the API's `DevicesController` altogether and to just call the `info(Request $request)` function directly for each implementation, but I couldn't figure out a graceful way to dynamically pick the correct instance of `IDeviceInformation` to call.