Fix for API's info method
There was an oversite where I was trying to use the currently authenticated user to retrieve info about a device.  This is problematic because the `info` method is called by the microcontroller which doesn't have the ability to authenticate users.  I reverted back to the old way where the desired user ID is passed as a query string parameter.  I had to create a `UserRepository` to make this work correctly.