Add Dusk tests to existing functionality
Added a suite of browser tests to test existing functionality.  This included fixing some shortcomings in the views to make the tests work (like adding `id`'s and `dusk` tags, fixing the label for the button when adding/editing a device so they weren't the same).  Also made the views more consistent with regard to `id` naming conventions I stumbled across.