Use User's model "ownsDevice" function instead of reimplementation
The `doesUserOwnDevice` function in the API's `DevicesController` is no longer needed and should have been removed a while ago in favor of the implementation of `ownsDevice` present in the `User` model.

Also created a custom exception handler to avoid having to use try-catch blocks to determine if a user exists or not.