Use reflection to call actions on devices
This makes it a lot simpler and cleaner to add new actions on existing devices and to add new types of devices.  The old way had a confusing `DeviceInformationServiceProvider` to handle picking which class to ask for device information.  The new way uses reflection to determine which class to ask for device information.